basic folder structure changed, converter to typescript
