+++ package.json
@@ -17,11 +17,10 @@
   "scripts": {
     "start": "node dist/index.js",
     "build": "tsc",
-    "dev": "nodemon src/index.ts"
+    "dev": "ts-node src/index.ts"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/express-session": "^1.17.4",
-    "nodemon": "^2.0.16"
+    "@types/express-session": "^1.17.4"
   }
 }
